+++ codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/protocols/parsers/XmlBindingTraitParserGenerator.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 
-package software.amazon.smithy.rust.codegen.smithy.protocols
+package software.amazon.smithy.rust.codegen.smithy.protocols.parsers
 
 import software.amazon.smithy.codegen.core.CodegenException
 import software.amazon.smithy.model.knowledge.HttpBinding
@@ -49,84 +49,6 @@ import software.amazon.smithy.rust.codegen.util.orNull
 import software.amazon.smithy.rust.codegen.util.outputShape
 import software.amazon.smithy.rust.codegen.util.toPascalCase
 import software.amazon.smithy.rust.codegen.util.toSnakeCase
-
-interface StructuredDataParserGenerator {
-    /**
-     * Generate a parse function for a given targeted as a payload.
-     * Entry point for payload-based parsing.
-     * Roughly:
-     * ```rust
-     * fn parse_my_struct(input: &[u8]) -> Result<MyStruct, XmlError> {
-     *      ...
-     * }
-     * ```
-     */
-    fun payloadParser(member: MemberShape): RuntimeType
-
-    /** Generate a parser for operation input
-     * Because only a subset of fields of the operation may be impacted by the document, a builder is passed
-     * through:
-     *
-     * ```rust
-     * fn parse_some_operation(inp: &[u8], builder: my_operation::Builder) -> Result<my_operation::Builder, XmlError> {
-     *   ...
-     * }
-     * ```
-     */
-    fun operationParser(operationShape: OperationShape): RuntimeType?
-
-    /**
-     * Because only a subset of fields of the operation may be impacted by the document, a builder is passed
-     * through:
-     *
-     * ```rust
-     * fn parse_some_error(inp: &[u8], builder: my_operation::Builder) -> Result<my_operation::Builder, XmlError> {
-     *   ...
-     * }
-     */
-    fun errorParser(errorShape: StructureShape): RuntimeType?
-
-    /**
-     * ```rust
-     * fn parse_document(inp: &[u8]) -> Result<Document, Error> {
-     *   ...
-     * }
-     * ```
-     */
-    fun documentParser(operationShape: OperationShape): RuntimeType
-}
-
-interface StructuredDataSerializerGenerator {
-    /**
-     * Generate a parse function for a given targeted as a payload.
-     * Entry point for payload-based parsing.
-     * Roughly:
-     * ```rust
-     * ```
-     */
-    fun payloadSerializer(member: MemberShape): RuntimeType
-
-    /** Generate a serializer for operation input
-     * Because only a subset of fields of the operation may be impacted by the document, a builder is passed
-     * through:
-     *
-     * ```rust
-     * fn parse_some_operation(inp: &[u8], builder: my_operation::Builder) -> Result<my_operation::Builder, XmlError> {
-     *   ...
-     * }
-     * ```
-     */
-    fun operationSeralizer(operationShape: OperationShape): RuntimeType?
-
-    /**
-     * ```rust
-     * fn parse_document(inp: &[u8]) -> Result<Document, Error> {
-     *   ...
-     * }
-     * ```
-     */
-    fun documentSerializer(): RuntimeType
-}
 
 class XmlBindingTraitParserGenerator(protocolConfig: ProtocolConfig, private val xmlErrors: RuntimeType) :
     StructuredDataParserGenerator {
